requirements.txt
fastapi
uvicorn

openai-whisper
torch
torchaudio
torchvision

pydub
python-multipart

streamlit
requests

python-jose
python-dotenv

sqlalchemy
pymysql
cryptography

tqdm

passlib[bcrypt]
pydantic[email]

redis